Isolate CreateClient and ConsensusState methods
