+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -1,3 +1,5 @@
 pub mod channel;
 pub mod connection;
+pub mod consensus_state;
+pub mod create_client;
 pub mod update_client;
